[embedded] Fix the type of the type descriminator parameter of swift_coroFrameAlloc on 32 bit platforms
The type of malloc_type_id_t is a 64bit integer (unsigned long long).

https://github.com/swiftlang/swift/issues/86069

+++ stdlib/public/core/EmbeddedRuntime.swift
@@ -150,7 +150,7 @@ func alignedAlloc(size: Int, alignment: Int) -> UnsafeMutableRawPointer? {
 }
 
 @c
-public func swift_coroFrameAlloc(_ size: Int, _ type: UInt) -> UnsafeMutableRawPointer? {
+public func swift_coroFrameAlloc(_ size: Int, _ type: UInt64) -> UnsafeMutableRawPointer? {
   return unsafe alignedAlloc(
     size: size,
     alignment: _swift_MinAllocationAlignment)
